unsigned long, to match machine's word size.

+++ BigInt.cpp
@@ -4,7 +4,7 @@
 
 #include "BigInt.h"
 
-BigInt::BigInt(uint8_t *mag, BigInt::Prime prime)
+BigInt::BigInt(unsigned long *mag, BigInt::Prime prime)
     : mag(mag), prime(prime),length(sizeof(mag)/mag[0])
 {
